bevy_egui::egui_node
Trait EguiBevyPaintCallbackImpl

pub trait EguiBevyPaintCallbackImpl: Send + Sync {
    // Required methods
    fn update(
        &self,
        info: PaintCallbackInfo,
        window_entity: RenderEntity,
        pipeline_key: EguiPipelineKey,
        world: &mut World,
    );
    fn render<'pass>(
        &self,
        info: PaintCallbackInfo,
        render_pass: &mut TrackedRenderPass<'pass>,
        window_entity: RenderEntity,
        pipeline_key: EguiPipelineKey,
        world: &'pass World,
    );

    // Provided method
    fn prepare_render<'w>(
        &self,
        info: PaintCallbackInfo,
        render_context: &mut RenderContext<'w>,
        window_entity: RenderEntity,
        pipeline_key: EguiPipelineKey,
        world: &'w World,
    ) { ... }
}

Callback that executes custom rendering logic

Required Methods

fn update( &self, info: PaintCallbackInfo, window_entity: RenderEntity, pipeline_key: EguiPipelineKey, world: &mut World, )

Paint callback will be rendered in near future, all data must be finalized for render step

fn render<'pass>( &self, info: PaintCallbackInfo, render_pass: &mut TrackedRenderPass<'pass>, window_entity: RenderEntity, pipeline_key: EguiPipelineKey, world: &'pass World, )

Paint callback render step

Native wgpu RenderPass can be retrieved from TrackedRenderPass by calling TrackedRenderPass::wgpu_pass.

Provided Methods

fn prepare_render<'w>( &self, info: PaintCallbackInfo, render_context: &mut RenderContext<'w>, window_entity: RenderEntity, pipeline_key: EguiPipelineKey, world: &'w World, )

Paint callback call before render step

Can be used to implement custom render passes or to submit command buffers for execution before egui render pass