Module bevy_egui::render_systems

Plugin systems for the render app.

Structs

EguiPipelines

Cached Pipeline IDs for the specialized EguiPipelines

EguiTextureBindGroups

Maps Egui textures to bind groups.

EguiTransform

Scale and translation for rendering Egui shapes. Is needed to transform Egui coordinates from the screen space with the center at (0, 0) to the normalised viewport space.

EguiTransforms

Describes the transform buffer.

ExtractedEguiSettings

Extracted Egui settings.

ExtractedEguiTextures

Extracted Egui textures.

Enums

EguiTextureId

Corresponds to Egui’s egui::TextureId.

Functions

extract_egui_render_data_system

Extracts Egui context, render output, settings and application window sizes.

extract_egui_textures_system

Extracts Egui textures.

prepare_egui_transforms_system

Prepares Egui transforms.

queue_bind_groups_system

Queues bind groups.

queue_pipelines_system

Queue EguiPipelines specialized on each window’s swap chain texture format.

setup_new_windows_render_system

Sets up the pipeline for newly created windows.