bevy_egui
Module render_systems

Plugin systems for the render app.

Structs

EguiPass

RenderLabel type for the Egui pass.

EguiPipelines

Cached Pipeline IDs for the specialized instances of EguiPipeline.

EguiTextureBindGroups

Maps Egui textures to bind groups.

EguiTransform

Scale and translation for rendering Egui shapes. Is needed to transform Egui coordinates from the screen space with the center at (0, 0) to the normalised viewport space.

EguiTransforms

Describes the transform buffer.

ExtractedEguiManagedTextures

The extracted version of EguiManagedTextures.

ExtractedEguiSettings

Extracted Egui settings.

ExtractedEguiTextures

Extracted Egui textures.

Enums

EguiTextureId

Corresponds to Egui’s egui::TextureId.

Functions

prepare_egui_transforms_system

Prepares Egui transforms.

queue_bind_groups_system

Queues bind groups.

queue_pipelines_system

Queue EguiPipeline instances specialized on each window’s swap chain texture format.

setup_new_render_to_image_nodes_system

Sets up render nodes for newly created “render to texture” Egui contexts.

setup_new_window_nodes_system

Sets up render nodes for newly created window Egui contexts.

teardown_render_to_image_nodes_system

Tears render nodes down for deleted “render to texture” Egui contexts.

teardown_window_nodes_system

Tears render nodes down for deleted window Egui contexts.